TT-7441 Color the racetrack using centralized theme
